Move `Options` to its own module
This is an internal module, so the API remains unchanged as
`textwrap::Options`.
